refac(tests): separate modules Split the overgrown `tests.py` into three separate modules in a new `tests` package.  Should make it a little bit easier to navigate the backend test suite.
