+++ backend/api/tests/__init__.py
@@ -1,0 +1,8 @@
+"""Django tests for the api module.
+Includes tests for the models, URLs and views.
+Intentionally excludes the following modules:
+ - __init__: who tests init?  Is that a thing people do?
+ - admin: not currently in use.
+ - apps: Django boilerplate code
+ - serializers: restframework boilerplace code
+"""
